Modify: 헤더, 푸터 컴포넌트 Layout 폴더 안으로 포함시킴 헤더와 푸터 각각에 대응되는 스타일은 CommonStyle에서 -> 각각의 스타일 파일로 이동
