+++ src/components/Shared/Layout/Footer/Footer.jsx
@@ -1,12 +1,12 @@
-import { Wrapper, FixedFooter } from "@/styles/CommonStyle";
+import { FooterWrapper, FixedFooter } from "./FooterStyle";
 
 const Footer = (props) => {
   let message = "KKUJJANG ⓒ All rights reserved.";
 
   return (
-    <Wrapper row="center" col="center">
+    <FooterWrapper row="center" col="center">
       <FixedFooter>{message}</FixedFooter>
-    </Wrapper>
+    </FooterWrapper>
   );
 };
