+++ src/components/Shared/Pagination/Pagination.jsx
@@ -13,15 +13,14 @@ import {
 const Pagination = ({ page, setPage, limit, lastPageIdx = 4 }) => {
   // 마지막 페이지 번호를 통해 배열 초기화
   const pageList = Array.from({ length: lastPageIdx }, (_, i) => i);
-  // const [currPageArr, setCurrPageArr] = useState(Array(lastPageIdx.fill(null)));
-  // const [pageList, setPageList] = useState(new Array(lastPageIdx));
-  // debugger;
+
   let tempString = "<<";
   useEffect(() => {}, [page]);
 
+  // 
   return (
     <PaginationWrapper row="center" col="center">
-      <FirstButton>{tempString}</FirstButton>
+      <FirstButton onClick={() => setPage(1)}>{tempString}</FirstButton>
       <PrevButton>{tempString}</PrevButton>
       <PageList>
         {pageList.map((i) => (
@@ -30,8 +29,8 @@ const Pagination = ({ page, setPage, limit, lastPageIdx = 4 }) => {
           </PageItem>
         ))}
       </PageList>
-      <NextButton>{tempString}</NextButton>
-      <LastButton>{tempString}</LastButton>
+      <NextButton onClick={() => setPage(lastPageIdx)}>{tempString}</NextButton>
+      <LastButton onClick={() => setPage(lastPageIdx)}>{tempString}</LastButton>
     </PaginationWrapper>
   );
 };
@@ -39,7 +38,7 @@ const Pagination = ({ page, setPage, limit, lastPageIdx = 4 }) => {
 Pagination.propTypes = {
   page: PropTypes.number, // 현재 페이지 번호
   setPage: PropTypes.func, // 현재 페이지 업데이트 함수
-  limit: PropTypes.number, // 한 페이지에 보여줄 게시물 개수
+  limit: PropTypes.number, // 
   lastPageIdx: PropTypes.number // 마지막 페이지 번호
 };
